+++ lib/bitstampede/Net.rb
